Colors, .gitignore
- First colors are now added in: Adding in colors from previous site
except for the Forest Green, which we dropped last year at some point.
All color values are being stored in a data file
(_data/stack-colors.yml). The page needs styling, but I’ll leave that
to @tgoas and co.
- Second adding in a .gitignore since we didn’t have one. This ignores
most of the basic system files we don’t need, but also the /_site/
folder.