Explicitly set the left margin in sidebar widget sub-navigation.
The original design spec had this flush left, but the default Q&A CSS for lists caused a 30px margin here, which is why uses of this pattern in Q&A code added explicit `ml8` or `ml12` to fix.

It seems that a little bit of margin is indeed warranted; 8px seems reasonable.