Remove font smoothing
This is really a tough call. As of macOS 11.3, the user can no longer choose their antialiasing at the OS level. This _feels_ like a bug or oversight, so the idea was we would be opinionated since the users can't make a change on their end anymore. Not everyone is on 11.3, and I hope this is fixed in 11.4, so let's revisit then.