Style readonly attribute (not read-only)
Addresses: https://github.com/StackExchange/Stacks/issues/620

The attribute to target is `readonly`. It can also be targeted as a pseudo-class in the form of `:read-only`, which is what I'd guess caused the confusion here.

Note: This targets a `readonly` attribute on `<select />` elements, though that element does not support it.
See: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/select#attributes