Add a notice that dark mode colors are inverted
See https://meta.stackexchange.com/questions/368241/please-document-the-inversion-of-colours-in-dark-mode-on-the-stacks-website.

A user was confused because they were comparing colors between the design site in dark mode and a Stack Exchange site without dark mode support, and were confused why the colors were apparently mismatched, when writing a custom style sheet. Add a note that in order to compare colors, the design site must be in the same mode as the SE site.