fix(breadcrumbs): meet WCAG 2 AA contrast
Note: contrast is judged in the context of the Stacks documentation. In this case, the breadcrumbs are within a container with a background color different from the main content background color. This may result in an overcompensation for color contrast. That said, we can look at this as a little extra reassurance that if the component is used on a different background color, it's still likely to pass accessibility checks.