update Stimulus to latest
Aside from the actual upgrade to Stimulus 3, this requires a few other
changes. Most importantly, TypeScript compilation can no longer target
ES5 because Stimulus dropped support for Internet Explorer. This
technically makes this a breaking change (but it should be the only
one; feature-wise, Stimulus 3 is backwards-compatible to Stimulus 2).

The switch to ES6 also had an effect on how imports are reordered during
transpilation, which required a minor change to the documentation code
to ensure that jQuery is available before global.search.js is executed.

Final note, just because I got confused by it myself initially: Stimulus
2 temporarily still exists as a (transitive) dev dependency because
Stacks depends on itself (via stacks-editor).