fix(tag): improve focus style to meet WCAG 2.2 AA
This commit also adds some minor styling updates to account for button-based `.s-tag--dismiss` elements.

These changes should result in no changes to existing `.s-tag--dismiss` elements, but the docs now display the semantically proper usage of a button element for the `X` button with the parent `.s-tag` being rendered as a span and consumers can now use a button element for the `.s-tag--dismiss` child without needing to apply reset classes.